services: websocketservice parsed to typescript
