+++ package.json
@@ -20,8 +20,7 @@
     "jsdom": "^17.0.0",
     "mongoose": "^6.0.0",
     "node-cron": "^3.0.0",
-    "socket.io": "^3.1.1",
-    "socket.io-client": "^3.1.1"
+    "socket.io": "^3.1.1"
   },
   "devDependencies": {
     "@babel/cli": "^7.18.6",
@@ -35,6 +34,7 @@
     "@types/jest": "^28.1.6",
     "@types/jsdom": "^16.2.15",
     "@types/node": "^18.0.3",
+    "@types/node-cron": "^3.0.2",
     "@types/supertest": "^2.0.12",
     "babel-plugin-inline-dotenv": "^1.7.0",
     "babel-plugin-module-resolver": "^4.1.0",
